Move the DecorationController into the hosts/vscode package.
