+++ src/hosts/vscode/DecorationController.ts
